+++ src/OpenApiWebApi/Program.cs
@@ -1,13 +1,36 @@
-namespace OpenApiWebApi;
+var builder = WebApplication.CreateBuilder(args);
+var env = builder.Environment;
+var services = builder.Services;
 
-public class Program {
-	public static void Main(string[] args) {
-		CreateHostBuilder(args).Build().Run();
-	}
+services.AddControllers();
 
-	public static IHostBuilder CreateHostBuilder(string[] args) =>
-		Host.CreateDefaultBuilder(args)
-			.ConfigureWebHostDefaults(webBuilder => {
-				webBuilder.UseStartup<Startup>();
-			});
+services.Configure<RouteOptions>(options => {
+	options.LowercaseQueryStrings = true;
+	options.LowercaseUrls = true;
+});
+
+// OpenAPI/Swagger
+services.AddOpenApiDocument(settings => {
+	settings.PostProcess = document => {
+		document.Info.Title = "Sample API";
+		document.Info.Description = "ASP.NET Core Web API";
+	};
+});
+
+var app = builder.Build();
+
+if (env.IsDevelopment()) {
+	app.UseDeveloperExceptionPage();
 }
+
+// OpenAPI/Swagger
+app.UseOpenApi(settings => {
+});
+app.UseSwaggerUi(settings => {
+});
+
+app.UseRouting();
+
+app.MapControllers();
+
+app.Run();
