+++ src/PolicyAuthzWebApp/PolicyNames.cs
@@ -1,0 +1,6 @@
+namespace PolicyAuthzWebApp;
+
+public static class PolicyNames {
+	public const string Authenticated = nameof(Authenticated);
+	public const string AdminRole = nameof(AdminRole);
+}
