+++ src/PolicyAuthzWebApp/Program.cs
@@ -1,13 +1,71 @@
-namespace PolicyAuthzWebApp;
+using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.AspNetCore.Authorization;
+using PolicyAuthzWebApp;
 
-public class Program {
-	public static void Main(string[] args) {
-		CreateHostBuilder(args).Build().Run();
-	}
+var builder = WebApplication.CreateBuilder(args);
+var config = builder.Configuration;
+var env = builder.Environment;
+var services = builder.Services;
 
-	public static IHostBuilder CreateHostBuilder(string[] args) =>
-		Host.CreateDefaultBuilder(args)
-			.ConfigureWebHostDefaults(webBuilder => {
-				webBuilder.UseStartup<Startup>();
-			});
+// 認証
+services
+	.AddAuthentication(options => {
+		options.DefaultScheme = CookieAuthenticationDefaults.AuthenticationScheme;
+	}).AddCookie(options => {
+	});
+
+// 承認
+services
+	.AddAuthorizationBuilder()
+	.AddPolicy(PolicyNames.Authenticated, builder => {
+		builder.RequireAuthenticatedUser();
+	})
+	.AddPolicy(PolicyNames.AdminRole, builder => {
+		builder.RequireRole("admin");
+	});
+// todo:
+/*
+	.AddPolicy("AreaRoles", builder => {
+		builder.RequireAreaRoles(area: "Admin", roles: "admin");
+	});
+*/
+
+// MVC
+services.AddControllers(options => {
+	// 承認（Authorization）は
+	// グローバルフィルタに指定するか
+	// コントローラにAuthorizeAttributeを指定するか
+	// エンドポイントにRequireAuthorizationするか
+	/*
+	options.Filters.Add(new AuthorizeFilter("Authenticated"));
+	*/
+});
+
+var app = builder.Build();
+
+if (env.IsDevelopment()) {
+	app.UseDeveloperExceptionPage();
 }
+
+app.UseRouting();
+
+// 認証
+app.UseAuthentication();
+
+// 承認
+// AuthorizeAttributeのために必要
+app.UseAuthorization();
+
+app.MapAreaControllerRoute(
+		name: "admin",
+		areaName: "Admin",
+		pattern: "Admin/{controller=Default}/{action=Index}/{id?}")
+	.RequireAuthorization(new AuthorizeAttribute(PolicyNames.AdminRole));
+
+app.MapControllerRoute(
+		name: "default",
+		pattern: "{controller=Default}/{action=Index}/{id?}")
+	// コントローラやアクションにAuthorizeAttributeを指定することをほぼ同じことだと思う
+	.RequireAuthorization(new AuthorizeAttribute(PolicyNames.Authenticated));
+
+app.Run();
